:sparkles: Add LTO support
LTO is default enabled in cmake-util in order to enable it for all
builds. It will NOT be enabled for the linker. This gives the developer
the option to turn on/off LTO at the linker stage.

When LTO is not used at the linker, all of the LTO information in the
archive and library files gets eliminated.